1.目标卡路里进制转换 2.只在修改体重、性别后需要重启追踪器 3.定位页宠物头像修改

+++ app/src/main/java/com/xiaomaoqiu/now/bussiness/pet/info/petdata/SportIndexActivity.java
@@ -95,7 +95,7 @@ public class SportIndexActivity extends BaseActivity implements IChartCallback {
         int totalWidth=threePartLineView_sport.getWidth();
         int targetWidth=edSport>0.0?(int)((edSport*totalWidth)/targetSport):totalWidth;
         textAimView_sport.setAim((int)edSport+"",(int)targetSport+"", targetWidth);
-        String tip="今日目标消耗为"+(int)targetSport+"千卡，实际消耗为"+(int)edSport+"千卡。";
+        String tip="今日目标消耗为"+targetSport+"千卡，实际消耗为"+edSport+"千卡。";
         todayTip.setText(tip);
     }
